@Column length 由 int 更改为 long

+++ database-core/src/main/java/gaarason/database/eloquent/annotations/Column.java
@@ -67,7 +67,7 @@ public @interface Column
      * (Optional) The column length. (Applies only if a
      * string-valued column is used.)
      */
-    int length() default 255;
+    long length() default 255;
 
     /**
      * (Optional) The precision for a decimal (exact numeric)
